+++ src/srv/ayats/ayat_id_by_sura_ayat_num.py
@@ -22,7 +22,6 @@
 
 # TODO #899 Перенести классы в отдельные файлы 41
 
-import uuid
 from typing import final, override
 
 import attrs
@@ -77,29 +76,3 @@ class AyatIdBySuraAyatNum(AsyncInt):
         return row['ayat_id']
 
 
-@final
-@attrs.define(frozen=True)
-@elegant
-class AyatIdByPublicId(AsyncInt):
-    """Поиск аятов по номеру суры, аята."""
-
-    _public_id: uuid.UUID
-    _pgsql: Database
-
-    @override
-    async def to_int(self) -> AyatId:
-        """Числовое представление.
-
-        :return: int
-        :raises AyatNotFoundError: если аят не найден
-        """
-        query = '\n'.join([
-            'SELECT ayat_id FROM ayats',
-            'WHERE public_id = :public_id',
-        ])
-        row = await self._pgsql.fetch_one(query, {
-            'public_id': str(self._public_id),
-        })
-        if not row:
-            raise AyatNotFoundError
-        return row['ayat_id']
